Fix getSubUser() to only return user with exact username match
Retrieving subusers is search as opposed to a specific retrieval,
meaning that the API query can return multiple users if the username is
contained within the username. Previously we returned the first user as
retrieved from the API, which could give us the wrong user if several
were matching.

This patch fixes the problem by doing a post-API strict match against
the usernames of the retrieved subusers.